requirements.txt
altair==5.3.0
attrs==23.2.0
blinker==1.8.2
branca==0.7.2
cachetools==5.3.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
et-xmlfile==1.1.0
folium==0.17.0
fonttools==4.57.0
geopandas==1.0.1
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
importlib_metadata==7.2.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.1
mdurl==0.1.2
narwhals==1.37.1
numpy==1.26.4
openpyxl==3.1.2
packaging==23.2
pandas==2.2.2
pillow==10.4.0
plotly==6.0.0
protobuf==4.25.3
pyarrow==16.1.0
pydeck==0.9.1
Pygments==2.18.0
pynasapower==0.0.1a1
pyogrio==0.9.0
pyparsing==3.2.3
pyproj==3.6.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.19.0
scipy==1.14.0
shapely==2.0.5
six==1.16.0
smmap==5.0.1
streamlit==1.50.0
streamlit-folium==0.18.0
tenacity==8.5.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
validators==0.32.0
watchdog==4.0.1
windrose==1.9.2
xarray==2024.6.0
xyzservices==2024.6.0
zipp==3.19.2
